Almost Ready to template!
